requirements.txt
Django==2.2
flake8==3.7.7
pylint==2.3.1
djangorestframework==3.9.2
django-filter==2.1.0
Markdown==3.1
requests==2.21.0
django-cors-headers==2.5.2
django-cron==0.5.1
geopy==1.19.0
python-dateutil==2.6.0